test: isolate managed catalog timing from coverage instrumentation
