+++ docs/testing/behavior-contracts.json
@@ -7235,11 +7235,14 @@
     "priority": "P0",
     "status": "automated",
     "owners": [
-      "tests/contract/projects/managedRemotePerformance.test.js"
+      "tests/contract/projects/managedRemotePerformance.test.js",
+      "tests/unit/tooling/ciContracts.test.js",
+      "scripts/run-managed-remote-performance-checks.js"
     ],
     "evidence": [
       "src/projects/managedRemote/merge.ts",
-      "src/projects/managedRemote/payload.ts"
+      "src/projects/managedRemote/payload.ts",
+      "tests/fixtures/managedRemoteCatalog.js"
     ]
   },
   {
