+++ package.json
@@ -613,11 +613,12 @@
         "test:release-notes": "node scripts/run-release-notes-checks.js",
         "test:release-packaging": "node scripts/seed-release-packaging-stale-output.js && npm run package:release && node scripts/run-release-packaging-checks.js",
         "test:conversation-performance": "node scripts/run-conversation-performance-checks.js",
+        "test:managed-remote-performance": "node scripts/run-managed-remote-performance-checks.js",
         "test:conversation-sources:remote": "node --test tests/platform/remote/conversationSources.test.js",
         "test:extension-host": "npm run package:release && node scripts/run-extension-host-tests.js",
-        "test:ci:linux": "npm run test-compile && npm run brand:verify && npm run brand:check && npm run test:behavior-contracts && npm run lint:ci && npm run test:coverage:run && npm run test:conversation-sources:remote && npm run test:conversation-performance && npm run test:browser:run && npm run test:safety:run && npm run test:dashboard:run && npm run test:architecture-baseline && npm run test:architecture-guards && npm run test:architecture-policy && npm run test:release-notes && npm run test:release-packaging && node scripts/check-coverage-baseline.js && node scripts/check-changed-coverage.js",
+        "test:ci:linux": "npm run test-compile && npm run brand:verify && npm run brand:check && npm run test:behavior-contracts && npm run lint:ci && npm run test:coverage:run && npm run test:conversation-sources:remote && npm run test:conversation-performance && npm run test:managed-remote-performance && npm run test:browser:run && npm run test:safety:run && npm run test:dashboard:run && npm run test:architecture-baseline && npm run test:architecture-guards && npm run test:architecture-policy && npm run test:release-notes && npm run test:release-packaging && node scripts/check-coverage-baseline.js && node scripts/check-changed-coverage.js",
         "test:ci:linux:core": "npm run test-compile && npm run lint:ci && npm run test:coverage:run && node scripts/check-coverage-baseline.js && node scripts/check-changed-coverage.js",
-        "test:ci:linux:browser": "npm run test-compile && npm run test:browser:run && npm run test:dashboard:run && npm run test:conversation-performance",
+        "test:ci:linux:browser": "npm run test-compile && npm run test:browser:run && npm run test:dashboard:run && npm run test:conversation-performance && npm run test:managed-remote-performance",
         "test:ci:linux:safety": "npm run test-compile && npm run test:conversation-sources:remote && npm run test:safety:run && npm run test:architecture-baseline && npm run test:architecture-guards && npm run test:architecture-policy",
         "test:ci:linux:release": "npm run test-compile && npm run brand:verify && npm run brand:check && npm run test:behavior-contracts && npm run test:release-notes && npm run test:release-packaging",
         "brand:check": "node scripts/check-brand-identity.js",
